* Version 4.3.0
* Exact changes are described in CHANGELOG.md
