+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: dotenv
-version: 4.2.0
+version: 4.3.0
 
 description: Load environment variables from a `.env` file.
 homepage: https://github.com/mockturtl/dotenv
@@ -9,6 +9,7 @@ environment:
 dependencies:
   args: ^2.3.0
   meta: ^1.7.0
+  universal_io: ^2.2.2
 
 dev_dependencies:
   tidy: ^3.1.0
